+++ ChangeLog.md
@@ -2,6 +2,8 @@
 
 2024-04
 
+- added `scripts/nrs11.py`, basically as in the batch notebook.
+  Can be run under this repo with: `poetry run python scripts/nrs11.py`
 - make sure loggers are independent of each other (one per processing day).
   Annoyingly, `loguru.Logger` type hint is not recognized when running in the command line,
   so commenting out that for now.